+++ src/lib/discovery/index.ts
@@ -2,5 +2,5 @@ export { getPredefinedBootstrapNodes } from "./predefined";
 export * as predefined from "./predefined";
 export { Bootstrap, BootstrapOptions } from "./bootstrap";
 export * as dns from "./dns";
-export { Endpoints, DnsOverHttps } from "./dns_over_https";
+export { DnsOverHttps } from "./dns_over_https";
 export { ENRTree, ENRTreeValues, ENRRootValues } from "./enrtree";