+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   the latter does not bring Buffer/Long deps, is ESM compatible and remove the need for bufbuild and protoc.
 - Move package to `"type": "module"`.
 - Use ESM code in Mocha and Karma tests.
+- Upgrade `dns-query` dependency, breaking change on `DnsNodeDiscovery` API.
 
 ## [0.24.0] - 2022-05-27